+++ src/heat.js
@@ -1542,21 +1542,22 @@
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    function exportAllData( bindingOptions ) {
+    function exportAllData( bindingOptions, exportType ) {
         var contents = null,
-            contentsMimeType = getExportMimeType( bindingOptions );
+            contentsMimeType = getExportMimeType( bindingOptions ),
+            contentExportType = isDefined( exportType ) ? exportType.toLowerCase() : bindingOptions.exportType.toLowerCase();
 
-        if ( bindingOptions.exportType.toLowerCase() === _export_Type_Csv ) {
+        if ( contentExportType === _export_Type_Csv ) {
             contents = getCsvContent( bindingOptions );
-        } else if ( bindingOptions.exportType.toLowerCase() === _export_Type_Json ) {
+        } else if ( contentExportType === _export_Type_Json ) {
             contents = getJsonContent( bindingOptions );
-        } else if ( bindingOptions.exportType.toLowerCase() === _export_Type_Xml ) {
+        } else if ( contentExportType === _export_Type_Xml ) {
             contents = getXmlContents( bindingOptions );
-        } else if ( bindingOptions.exportType.toLowerCase() === _export_Type_Txt ) {
+        } else if ( contentExportType === _export_Type_Txt ) {
             contents = getTxtContents( bindingOptions );
         }
 
-        if ( contents !== _string.empty ) {
+        if ( isDefinedString( contents ) ) {
             var tempLink = createElement( _parameter_Document.body, "a" );
             tempLink.style.display = "none";
             tempLink.setAttribute( "target", "_blank" );
@@ -2607,12 +2608,13 @@
      * @fires       onExport
      * 
      * @param       {string}    elementId                                   The Heat.js element ID whose data should be exported.
+     * @param       {string}    [exportType]                                The export type to use (defaults to "csv", also accepts "json", "xml", and "txt").
      * 
      * @returns     {Object}                                                The Heat.js class instance.
      */
-    this.export = function( elementId ) {
+    this.export = function( elementId, exportType ) {
         if ( isDefinedString( elementId ) && _elements_DateCounts.hasOwnProperty( elementId ) ) {
-            exportAllData( _elements_DateCounts[ elementId ].options );
+            exportAllData( _elements_DateCounts[ elementId ].options, exportType );
         }
 
         return this;